Chapter02_Python_Tools/requirements.txt
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.6.0
torch>=2.0.0
tensorflow>=2.10.0
paddlepaddle>=2.4.0
notebook